refactor(services): simplify service list generation and extract logic into services-data module
